Fix undefined error arising when calling the `gotData` function
The parameter `result` is undefined when the function is called in the current way. Update to use an arrow function (could also use `,function(err,result)` as documented in ml5's [lstm.generate](https://ml5js.org/docs/LSTMGenerator) documentation. Also, note that what we display is not `result.generated`, but just `result` (result is not an object, just a string).